requirements.txt
numpy==1.22.3
Pillow==9.1.0
streamlit==1.9.0
tensorflow==2.8.0
tensorflow_hub==0.12.0
